refactor: split loop and schema modules
