+++ src/agent/loop_.rs
@@ -2717,7 +2717,6 @@ mod tests {
     use std::collections::VecDeque;
     use std::sync::atomic::{AtomicUsize, Ordering};
     use std::sync::{Arc, Mutex};
-    use std::time::Duration;
 
     #[test]
     fn scrub_credentials_redacts_bearer_token() {
@@ -2919,125 +2918,6 @@ mod tests {
             Ok(crate::tools::ToolResult {
                 success: true,
                 output: format!("counted:{value}"),
-                error: None,
-            })
-        }
-    }
-
-    struct RecordingArgsTool {
-        name: String,
-        recorded_args: Arc<Mutex<Vec<serde_json::Value>>>,
-    }
-
-    impl RecordingArgsTool {
-        fn new(name: &str, recorded_args: Arc<Mutex<Vec<serde_json::Value>>>) -> Self {
-            Self {
-                name: name.to_string(),
-                recorded_args,
-            }
-        }
-    }
-
-    #[async_trait]
-    impl Tool for RecordingArgsTool {
-        fn name(&self) -> &str {
-            &self.name
-        }
-
-        fn description(&self) -> &str {
-            "Records tool arguments for regression tests"
-        }
-
-        fn parameters_schema(&self) -> serde_json::Value {
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "prompt": { "type": "string" },
-                    "schedule": { "type": "object" },
-                    "delivery": { "type": "object" }
-                }
-            })
-        }
-
-        async fn execute(
-            &self,
-            args: serde_json::Value,
-        ) -> anyhow::Result<crate::tools::ToolResult> {
-            self.recorded_args
-                .lock()
-                .expect("recorded args lock should be valid")
-                .push(args.clone());
-            Ok(crate::tools::ToolResult {
-                success: true,
-                output: args.to_string(),
-                error: None,
-            })
-        }
-    }
-
-    struct DelayTool {
-        name: String,
-        delay_ms: u64,
-        active: Arc<AtomicUsize>,
-        max_active: Arc<AtomicUsize>,
-    }
-
-    impl DelayTool {
-        fn new(
-            name: &str,
-            delay_ms: u64,
-            active: Arc<AtomicUsize>,
-            max_active: Arc<AtomicUsize>,
-        ) -> Self {
-            Self {
-                name: name.to_string(),
-                delay_ms,
-                active,
-                max_active,
-            }
-        }
-    }
-
-    #[async_trait]
-    impl Tool for DelayTool {
-        fn name(&self) -> &str {
-            &self.name
-        }
-
-        fn description(&self) -> &str {
-            "Delay tool for testing parallel tool execution"
-        }
-
-        fn parameters_schema(&self) -> serde_json::Value {
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "value": { "type": "string" }
-                },
-                "required": ["value"]
-            })
-        }
-
-        async fn execute(
-            &self,
-            args: serde_json::Value,
-        ) -> anyhow::Result<crate::tools::ToolResult> {
-            let now_active = self.active.fetch_add(1, Ordering::SeqCst) + 1;
-            self.max_active.fetch_max(now_active, Ordering::SeqCst);
-
-            tokio::time::sleep(Duration::from_millis(self.delay_ms)).await;
-
-            self.active.fetch_sub(1, Ordering::SeqCst);
-
-            let value = args
-                .get("value")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or_default()
-                .to_string();
-
-            Ok(crate::tools::ToolResult {
-                success: true,
-                output: format!("ok:{value}"),
                 error: None,
             })
         }
@@ -3274,215 +3154,6 @@ mod tests {
     }
 
     // should_execute_tools_in_parallel tests moved to tool_execution.rs
-
-    #[tokio::test]
-    async fn run_tool_call_loop_executes_multiple_tools_with_ordered_results() {
-        let provider = ScriptedProvider::from_text_responses(vec![
-            r#"<tool_call>
-{"name":"delay_a","arguments":{"value":"A"}}
-</tool_call>
-<tool_call>
-{"name":"delay_b","arguments":{"value":"B"}}
-</tool_call>"#,
-            "done",
-        ]);
-
-        let active = Arc::new(AtomicUsize::new(0));
-        let max_active = Arc::new(AtomicUsize::new(0));
-        let tools_registry: Vec<Box<dyn Tool>> = vec![
-            Box::new(DelayTool::new(
-                "delay_a",
-                200,
-                Arc::clone(&active),
-                Arc::clone(&max_active),
-            )),
-            Box::new(DelayTool::new(
-                "delay_b",
-                200,
-                Arc::clone(&active),
-                Arc::clone(&max_active),
-            )),
-        ];
-
-        let approval_cfg = crate::config::AutonomyConfig {
-            level: crate::security::AutonomyLevel::Full,
-            ..crate::config::AutonomyConfig::default()
-        };
-        let approval_mgr = ApprovalManager::from_config(&approval_cfg);
-
-        let mut history = vec![
-            ChatMessage::system("test-system"),
-            ChatMessage::user("run tool calls"),
-        ];
-        let observer = NoopObserver;
-
-        let result = run_tool_call_loop(
-            &provider,
-            &mut history,
-            &tools_registry,
-            &observer,
-            "mock-provider",
-            "mock-model",
-            0.0,
-            true,
-            Some(&approval_mgr),
-            "telegram",
-            None,
-            &crate::config::MultimodalConfig::default(),
-            4,
-            None,
-            None,
-            None,
-            &[],
-            &[],
-            None,
-            None,
-            &crate::config::PacingConfig::default(),
-        )
-        .await
-        .expect("parallel execution should complete");
-
-        assert_eq!(result, "done");
-        assert!(
-            max_active.load(Ordering::SeqCst) >= 1,
-            "tools should execute successfully"
-        );
-
-        let tool_results_message = history
-            .iter()
-            .find(|msg| msg.role == "user" && msg.content.starts_with("[Tool results]"))
-            .expect("tool results message should be present");
-        let idx_a = tool_results_message
-            .content
-            .find("name=\"delay_a\"")
-            .expect("delay_a result should be present");
-        let idx_b = tool_results_message
-            .content
-            .find("name=\"delay_b\"")
-            .expect("delay_b result should be present");
-        assert!(
-            idx_a < idx_b,
-            "tool results should preserve input order for tool call mapping"
-        );
-    }
-
-    #[tokio::test]
-    async fn run_tool_call_loop_injects_channel_delivery_defaults_for_cron_add() {
-        let provider = ScriptedProvider::from_text_responses(vec![
-            r#"<tool_call>
-{"name":"cron_add","arguments":{"job_type":"agent","prompt":"remind me later","schedule":{"kind":"every","every_ms":60000}}}
-</tool_call>"#,
-            "done",
-        ]);
-
-        let recorded_args = Arc::new(Mutex::new(Vec::new()));
-        let tools_registry: Vec<Box<dyn Tool>> = vec![Box::new(RecordingArgsTool::new(
-            "cron_add",
-            Arc::clone(&recorded_args),
-        ))];
-
-        let mut history = vec![
-            ChatMessage::system("test-system"),
-            ChatMessage::user("schedule a reminder"),
-        ];
-        let observer = NoopObserver;
-
-        let result = run_tool_call_loop(
-            &provider,
-            &mut history,
-            &tools_registry,
-            &observer,
-            "mock-provider",
-            "mock-model",
-            0.0,
-            true,
-            None,
-            "telegram",
-            Some("chat-42"),
-            &crate::config::MultimodalConfig::default(),
-            4,
-            None,
-            None,
-            None,
-            &[],
-            &[],
-            None,
-            None,
-            &crate::config::PacingConfig::default(),
-        )
-        .await
-        .expect("cron_add delivery defaults should be injected");
-
-        assert_eq!(result, "done");
-
-        let recorded = recorded_args
-            .lock()
-            .expect("recorded args lock should be valid");
-        let delivery = recorded[0]["delivery"].clone();
-        assert_eq!(
-            delivery,
-            serde_json::json!({
-                "mode": "announce",
-                "channel": "telegram",
-                "to": "chat-42",
-            })
-        );
-    }
-
-    #[tokio::test]
-    async fn run_tool_call_loop_preserves_explicit_cron_delivery_none() {
-        let provider = ScriptedProvider::from_text_responses(vec![
-            r#"<tool_call>
-{"name":"cron_add","arguments":{"job_type":"agent","prompt":"run silently","schedule":{"kind":"every","every_ms":60000},"delivery":{"mode":"none"}}}
-</tool_call>"#,
-            "done",
-        ]);
-
-        let recorded_args = Arc::new(Mutex::new(Vec::new()));
-        let tools_registry: Vec<Box<dyn Tool>> = vec![Box::new(RecordingArgsTool::new(
-            "cron_add",
-            Arc::clone(&recorded_args),
-        ))];
-
-        let mut history = vec![
-            ChatMessage::system("test-system"),
-            ChatMessage::user("schedule a quiet cron job"),
-        ];
-        let observer = NoopObserver;
-
-        let result = run_tool_call_loop(
-            &provider,
-            &mut history,
-            &tools_registry,
-            &observer,
-            "mock-provider",
-            "mock-model",
-            0.0,
-            true,
-            None,
-            "telegram",
-            Some("chat-42"),
-            &crate::config::MultimodalConfig::default(),
-            4,
-            None,
-            None,
-            None,
-            &[],
-            &[],
-            None,
-            None,
-            &crate::config::PacingConfig::default(),
-        )
-        .await
-        .expect("explicit delivery mode should be preserved");
-
-        assert_eq!(result, "done");
-
-        let recorded = recorded_args
-            .lock()
-            .expect("recorded args lock should be valid");
-        assert_eq!(recorded[0]["delivery"], serde_json::json!({"mode": "none"}));
-    }
 
     #[tokio::test]
     async fn run_tool_call_loop_deduplicates_repeated_tool_calls() {
